+++ craft_ai/client.py
@@ -37,6 +37,8 @@ ERROR_ID_MESSAGE = (
     "and must be between 1 and 36 characters."
 )
 
+ERROR_EMPTY_PAYLOAD = "Invalid payload given, it should contains a least one agent"
+
 
 def current_time_ms():
     return int(round(time.time() * 1000))
@@ -1186,6 +1188,9 @@ class Client(object):
                         valid_entity_indices.append(index)
                 else:
                     valid_entity_indices.append(index)
+
+        if len(payload) == 0:
+            raise CraftAiBadRequestError(ERROR_EMPTY_PAYLOAD)
 
         if len(invalid_entity_indices) == len(payload):
             raise CraftAiBadRequestError(ERROR_ID_MESSAGE)